corrige função e adiciona docstring
